refactor(app): MainWindow のクラッシュレポート機能を CrashReportDialogService へ分離 (#206)
MainWindow.xaml.cs (770行) に直接実装されていたクラッシュレポート
ダイアログ表示・GitHub Issue起動・メール起動・ログフォルダを開く処理を
専用サービスへ移動し、MainWindow肥大化防止ガードレール違反(イベント
ハンドラに処理本体が直接実装されていた)を是正。590行(-23%)に縮退。

GitHub Issue URL/mailto URI構築ロジックを純粋関数として抽出し、
単体テストを追加(#150 Phase4)。
