+++ PhotoGeoExplorer/MainWindow.xaml.cs
@@ -37,6 +37,7 @@ public sealed partial class MainWindow : Window, IDisposable
     private bool _windowIconSet;
     private readonly HelpService _helpService;
     private readonly MainWindowLayoutCoordinator _layoutCoordinator;
+    private readonly CrashReportDialogService _crashReportDialogService;
     internal FileBrowserPaneViewModel FileBrowserPaneViewModel => _fileBrowserPaneViewModel;
 
     internal void SetCrashReportService(Services.ICrashReportService crashReportService)
@@ -81,6 +82,11 @@ public sealed partial class MainWindow : Window, IDisposable
             value => _settingsCoordinator.ShowQuickStartOnStartup = value,
             () => _settingsCoordinator.SaveAsync(),
             _settingsCoordinator.SettingsFileExistsAtStartup,
+            (message, severity) => _viewModel.ShowNotificationMessage(message, severity));
+        _crashReportDialogService = new CrashReportDialogService(
+            dialogService,
+            () => _crashReportService,
+            () => _viewModel.CrashReportsDirectoryPath,
             (message, severity) => _viewModel.ShowNotificationMessage(message, severity));
         _viewModel.ConfigureHelpService(_helpService);
         _viewModel.ConfigureSettingsCoordinator(_settingsCoordinator);
@@ -449,190 +455,12 @@ public sealed partial class MainWindow : Window, IDisposable
 
     private async void OnReportCrashClicked(object sender, RoutedEventArgs e)
     {
-        var logContent = _crashReportService?.GetLatestCrashLogContent();
-
-        var summaryPanel = BuildCrashReportSummaryPanel(logContent);
-
-        var dialog = new ContentDialog
-        {
-            Title = LocalizationService.GetString("CrashReportDialog.Title"),
-            Content = summaryPanel,
-            PrimaryButtonText = LocalizationService.GetString("CrashReportDialog.GitHubButton"),
-            SecondaryButtonText = LocalizationService.GetString("CrashReportDialog.CopyButton"),
-            CloseButtonText = LocalizationService.GetString("CrashReportDialog.CloseButton"),
-            XamlRoot = Content.XamlRoot
-        };
-
-        var result = await dialog.ShowAsync();
-
-        if (result == ContentDialogResult.Primary)
-        {
-            await OpenCrashReportGitHubIssueAsync(logContent).ConfigureAwait(true);
-        }
-        else if (result == ContentDialogResult.Secondary)
-        {
-            await CopyLogAndOpenMailAsync(logContent).ConfigureAwait(true);
-        }
-    }
-
-    private StackPanel BuildCrashReportSummaryPanel(string? logContent)
-    {
-        var panel = new StackPanel { Spacing = 8, MaxWidth = 480 };
-
-        var version = ParseCrashLogField(logContent, "App Version:");
-        var timestamp = ParseCrashLogField(logContent, "Timestamp:");
-        var exType = ParseCrashLogField(logContent, "Exception Type:");
-
-        if (!string.IsNullOrEmpty(version) || !string.IsNullOrEmpty(timestamp) || !string.IsNullOrEmpty(exType))
-        {
-            var infoLines = new StackPanel { Spacing = 2 };
-            if (!string.IsNullOrEmpty(version))
-                infoLines.Children.Add(new TextBlock { Text = $"{LocalizationService.GetString("CrashReportDialog.LabelVersion")} {version}" });
-            if (!string.IsNullOrEmpty(timestamp))
-                infoLines.Children.Add(new TextBlock { Text = $"{LocalizationService.GetString("CrashReportDialog.LabelTimestamp")} {timestamp}" });
-            if (!string.IsNullOrEmpty(exType))
-                infoLines.Children.Add(new TextBlock { Text = $"{LocalizationService.GetString("CrashReportDialog.LabelException")} {exType}", TextWrapping = TextWrapping.Wrap });
-            panel.Children.Add(infoLines);
-        }
-
-        panel.Children.Add(new TextBlock
-        {
-            Text = LocalizationService.GetString("CrashReportDialog.SupportNote"),
-            TextWrapping = TextWrapping.Wrap,
-            Opacity = 0.7,
-            FontSize = 12
-        });
-
-        var folderLink = new HyperlinkButton
-        {
-            Content = LocalizationService.GetString("CrashReportDialog.OpenFolderLink"),
-            Padding = new Microsoft.UI.Xaml.Thickness(0)
-        };
-        folderLink.Click += OnOpenCrashReportFolderClicked;
-        panel.Children.Add(folderLink);
-
-        return panel;
-    }
-
-    private static string? ParseCrashLogField(string? logContent, string fieldName)
-    {
-        if (string.IsNullOrEmpty(logContent)) return null;
-        foreach (var line in logContent.Split('\n'))
-        {
-            if (line.StartsWith(fieldName, StringComparison.Ordinal))
-                return line[fieldName.Length..].Trim();
-        }
-        return null;
-    }
-
-    private static async Task OpenCrashReportGitHubIssueAsync(string? logContent)
-    {
-        const string baseUrl = "https://github.com/scottlz0310/PhotoGeoExplorer/issues/new";
-        var exType = ParseCrashLogField(logContent, "Exception Type:") ?? "Unknown";
-        var title = Uri.EscapeDataString($"[Problem] {exType}");
-
-        var truncated = logContent is { Length: > 2000 }
-            ? logContent[..2000] + "\n...(truncated)"
-            : logContent ?? "(no log)";
-
-        var body = Uri.EscapeDataString(
-            "## 問題レポート\n\n" +
-            "PhotoGeoExplorer の実行中に問題が検出されました。\n\n" +
-            "<details>\n<summary>診断ログ</summary>\n\n```\n" +
-            truncated +
-            "\n```\n\n</details>\n\n" +
-            "---\n*PhotoGeoExplorer から自動生成されました。*");
-
-        var url = $"{baseUrl}?title={title}&labels=bug&body={body}";
-        _ = await Windows.System.Launcher.LaunchUriAsync(new Uri(url)).AsTask().ConfigureAwait(true);
-    }
-
-    private static async Task CopyLogAndOpenMailAsync(string? logContent)
-    {
-        var text = logContent ?? LocalizationService.GetString("CrashReportDialog.NoLog");
-        var dataPackage = new Windows.ApplicationModel.DataTransfer.DataPackage();
-        dataPackage.SetText(text);
-        Windows.ApplicationModel.DataTransfer.Clipboard.SetContent(dataPackage);
-
-        var version = ParseCrashLogField(logContent, "App Version:") ?? string.Empty;
-        var exType = ParseCrashLogField(logContent, "Exception Type:") ?? "Unknown";
-        var subject = Uri.EscapeDataString($"[Problem Report] v{version} {exType}");
-        var body = Uri.EscapeDataString(LocalizationService.GetString("CrashReportDialog.MailBody"));
-        var mailto = new Uri($"mailto:photogeoexplorer@outlook.com?subject={subject}&body={body}");
-        _ = await Windows.System.Launcher.LaunchUriAsync(mailto).AsTask().ConfigureAwait(true);
-    }
-
-    private async void OnOpenCrashReportFolderClicked(object sender, RoutedEventArgs e)
-    {
-        var path = _viewModel.CrashReportsDirectoryPath;
-        if (string.IsNullOrWhiteSpace(path))
-        {
-            return;
-        }
-
-        try
-        {
-            if (!Directory.Exists(path))
-            {
-                Directory.CreateDirectory(path);
-            }
-
-            _ = await Windows.System.Launcher.LaunchFolderPathAsync(path);
-        }
-        catch (UnauthorizedAccessException ex)
-        {
-            HandleOpenLogFolderFailure(ex);
-        }
-        catch (IOException ex)
-        {
-            HandleOpenLogFolderFailure(ex);
-        }
-        catch (ArgumentException ex)
-        {
-            HandleOpenLogFolderFailure(ex);
-        }
-        catch (InvalidOperationException ex)
-        {
-            HandleOpenLogFolderFailure(ex);
-        }
+        await _crashReportDialogService.ShowCrashReportDialogAsync().ConfigureAwait(true);
     }
 
     private async void OnOpenLogFolderClicked(object sender, RoutedEventArgs e)
     {
-        try
-        {
-            var logDirectory = Path.GetDirectoryName(AppLog.LogFilePath);
-            if (string.IsNullOrWhiteSpace(logDirectory))
-            {
-                AppLog.Error("Log directory path is null or empty");
-                return;
-            }
-
-            if (!Directory.Exists(logDirectory))
-            {
-                Directory.CreateDirectory(logDirectory);
-                AppLog.Info($"Created log directory: {logDirectory}");
-            }
-
-            _ = await Windows.System.Launcher.LaunchFolderPathAsync(logDirectory);
-            AppLog.Info($"Opened log folder: {logDirectory}");
-        }
-        catch (UnauthorizedAccessException ex)
-        {
-            HandleOpenLogFolderFailure(ex);
-        }
-        catch (IOException ex)
-        {
-            HandleOpenLogFolderFailure(ex);
-        }
-        catch (ArgumentException ex)
-        {
-            HandleOpenLogFolderFailure(ex);
-        }
-        catch (InvalidOperationException ex)
-        {
-            HandleOpenLogFolderFailure(ex);
-        }
+        await _crashReportDialogService.OpenLogFolderAsync().ConfigureAwait(true);
     }
 
     private async void OnCheckUpdatesClicked(object sender, RoutedEventArgs e)
@@ -673,14 +501,6 @@ public sealed partial class MainWindow : Window, IDisposable
         {
             await HandleUpdateCheckFailureAsync(ex).ConfigureAwait(true);
         }
-    }
-
-    private void HandleOpenLogFolderFailure(Exception ex)
-    {
-        AppLog.Error("Failed to open log folder", ex);
-        _viewModel.ShowNotificationMessage(
-            LocalizationService.GetString("Message.FailedOpenLogFolder"),
-            InfoBarSeverity.Error);
     }
 
     private async Task HandleUpdateCheckFailureAsync(Exception ex)
