+++ PhotoGeoExplorer/Services/ICrashReportDialogService.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+
+namespace PhotoGeoExplorer.Services;
+
+internal interface ICrashReportDialogService
+{
+    Task ShowCrashReportDialogAsync();
+
+    Task OpenLogFolderAsync();
+}
